chore(util): Convert to TypeScript
