+++ apps/juxtaposition-ui/src/types/common/param-pack.ts
@@ -1,0 +1,16 @@
+export interface ParamPack {
+	title_id: string;
+	access_key: string;
+	platform_id: string;
+	region_id: string;
+	language_id: string;
+	country_id: string;
+	area_id: string;
+	network_restriction: string;
+	friend_restriction: string;
+	rating_restriction: string;
+	rating_organization: string;
+	transferable_id: string;
+	tz_name: string;
+	utc_offset: string;
+}
