+++ apps/juxtaposition-ui/src/types/common/token.ts
@@ -1,0 +1,8 @@
+export interface Token {
+	system_type: number;
+	token_type: number;
+	pid: number;
+	access_level: number;
+	title_id: bigint;
+	expire_time: bigint;
+}
