+++ apps/juxtaposition-ui/src/types/juxt/notification.ts
@@ -1,0 +1,9 @@
+export type Notification = {
+	pid: number;
+	type: 'notice' | 'yeah' | 'reply' | 'follow';
+	text: string;
+	image: string;
+	link: string;
+	objectID: string;
+	userPID: number;
+};
